fw update: always work with 4 byte chunks

+++ src/board/Board.h
@@ -433,7 +433,7 @@ namespace Board
         bool     isHWtriggerActive();
         uint32_t pageSize(size_t index);
         void     erasePage(size_t index);
-        void     fillPage(size_t index, uint32_t address, uint16_t value);
+        void     fillPage(size_t index, uint32_t address, uint32_t value);
         void     writePage(size_t index);
 #ifdef FW_BOOT
         // don't allow this API from application
